:sparkles: feature(none): 重构管理后台用户登陆注册流程，规划service 目录

+++ database/model/base.go
@@ -18,7 +18,7 @@ import (
 //      gorm.Model
 //    }
 type BaseModel struct {
-	ID        uint           `gorm:"primarykey" json:"id"`
+	ID        int64          `gorm:"primarykey" json:"id"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"deleted_at"`
